minor fixes for displaying quantity value and disable browser autocomplete for city fields

+++ src/main/java/com/ajurasz/model/StateHistory.java
@@ -41,17 +41,17 @@ public class StateHistory extends BaseEntity {
     public void setValue(BigDecimal value, boolean returned) {
         this.value = value;
         if(this.value.compareTo(new BigDecimal(0)) == 1) {
-            this.description = HistoryType.ADDED + " : " + value + " kg";
+            this.description = HistoryType.ADDED + " : " + value.intValue() + " kg";
         }
         if(this.value.compareTo(new BigDecimal(0)) == 0) {
-            this.description = HistoryType.ADDED + " : " + value + " kg";
+            this.description = HistoryType.ADDED + " : " + value.intValue() + " kg";
         }
         if(this.value.compareTo(new BigDecimal(0)) == -1) {
-            this.description = HistoryType.REMOVED + " : " + value.abs() + " kg";
+            this.description = HistoryType.REMOVED + " : " + value.abs().intValue() + " kg";
         }
 
         if(returned) {
-            this.description =  HistoryType.RETURNED + " : " + value.abs() + " kg";
+            this.description =  HistoryType.RETURNED + " : " + value.abs().intValue() + " kg";
         }
     }
